refactor authorization on api side
