+++ hms-api/src/handler/allPassAuthorizer.ts
@@ -1,0 +1,5 @@
+import { generatePolicy } from "./lambdaPolicyGenerator";
+
+export const alwaysPass = (event, context, callback) => {
+  callback(null, generatePolicy('user', 'Allow', event.methodArn));
+}
