+++ hms-api/src/handler/authorize.ts
@@ -1,59 +1,14 @@
-import {verify} from 'jsonwebtoken';
+import { alwaysPass } from "./allPassAuthorizer";
+import { authorizeWithActiveDirectory } from "./activeDirectoryAuthorizer";
 
-const AZURE_AD_CLIENT_ID = process.env.AZURE_AD_CLIENT_ID;
-const AZURE_AD_ISSUER = process.env.AZURE_AD_ISSUER;
-const AZURE_AD_CLIENT_PUBLIC_KEY = process.env.AZURE_AD_CLIENT_PUBLIC_KEY;
-
-const generatePolicy = (principalId, effect, resource) => {
-  const authResponse: any = {};
-  authResponse.principalId = principalId;
-  if (effect && resource) {
-    const policyDocument: any = {};
-    policyDocument.Version = '2012-10-17';
-    policyDocument.Statement = [];
-    const statementOne: any = {};
-    statementOne.Action = 'execute-api:Invoke';
-    statementOne.Effect = effect;
-    statementOne.Resource = resource;
-    policyDocument.Statement[0] = statementOne;
-    authResponse.policyDocument = policyDocument;
-  }
-  return authResponse;
-};
+const AUTHORIZER_TYPE = process.env.AUTHORIZER_TYPE || 'ALWAYS_PASS';
 
 export function authorize(event, context, callback) {
-  if (!event.authorizationToken) {
-    return callback('Unauthorized');
-  }
-
-  const tokenParts = event.authorizationToken.split(' ');
-  const tokenValue = tokenParts[1];
-
-  if (!(tokenParts[0].toLowerCase() === 'bearer' && tokenValue)) {
-    return callback('Unauthorized');
-  }
-  const options = {
-    audience: AZURE_AD_CLIENT_ID,
-    issuer: AZURE_AD_ISSUER,
-  };
-
-  try {
-    verify(
-      tokenValue,
-      AZURE_AD_CLIENT_PUBLIC_KEY,
-      options,
-      (verifyError, decoded) => {
-        if (verifyError) {
-          console.log(`Token invalid. ${verifyError}`);
-          return callback('Unauthorized');
-        }
-        return callback(
-          null,
-          generatePolicy(decoded.sub, 'Allow', event.methodArn)
-        );
-      }
-    );
-  } catch (err) {
+  if (AUTHORIZER_TYPE === 'ALWAYS_PASS') {
+    alwaysPass(event, context, callback);
+  } else if (AUTHORIZER_TYPE === 'ACTIVE_DIRECTORY') {
+    authorizeWithActiveDirectory(event, context, callback);
+  } else {
     return callback('Unauthorized');
   }
 }
