+++ hms-api/src/handler/lambdaPolicyGenerator.ts
@@ -1,0 +1,18 @@
+
+export const generatePolicy = (principalId, effect, resource) => {
+  const authResponse: any = {};
+  authResponse.principalId = principalId;
+  if (effect && resource) {
+    const policyDocument: any = {};
+    policyDocument.Version = '2012-10-17';
+    policyDocument.Statement = [];
+    const statementOne: any = {};
+    statementOne.Action = 'execute-api:Invoke';
+    statementOne.Effect = effect;
+    statementOne.Resource = resource;
+    policyDocument.Statement[0] = statementOne;
+    authResponse.policyDocument = policyDocument;
+  }
+  return authResponse;
+};
+
